requirements.txt
streamlit
openai
sentence-transformers
endee
pypdf
python-docx
